third_party/WebKit/LayoutTests/crypto/aes-ctr-parseAlgorithm-failures-expected.txt

Tests bad algorithm inputs for AES-CTR

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: SyntaxError: AesCtrParams: counter: Not an ArrayBufferView
error is: SyntaxError: AesCtrParams: counter: Missing required property
error is: SyntaxError: AesCtrParams: length: Missing required property
error is: SyntaxError: AesCtrParams: length: Outside of numeric range
error is: SyntaxError: AesCtrParams: length: Outside of numeric range
error is: SyntaxError: AesCtrParams: length: Outside of numeric range
PASS successfullyParsed is true

TEST COMPLETE
